+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 CHANGELOG
 =====
 
+Next Release
+-----
+
+* Daisy: allow for debug printing (off by default, increases program size due to formatting)
+* Daisy: set bootloader type in Makefile
+* DPF bugfixes: broken midi template include; MIDI_RT_CLOCK fails under certain conditions
+
 0.8.0
 -----
 
 * DPF: enumerated parameters
 * DPF: special `__hv_dpf_bpm` receiver of transport BPM value
-* DPF bugfixes: broken midi template include; MIDI_RT_CLOCK fails under certain conditions
 * Pdext: migrate to `pd-lib-builder` and newer `m_pd.h` and add some initial documentation.
 * deprecated: Fabric generator - no longer supported
 * docs: updates on missing objects and limitations