+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =====
 
+Next Release
+-----
+
+* Daisy: allow for debug printing (off by default, increases program size due to formatting)
+* Daisy: set bootloader type in Makefile
+* DPF bugfixes: broken midi template include; MIDI_RT_CLOCK fails under certain conditions
+
 0.8.0
 -----