+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Next Release
 
 * Daisy: allow for debug printing (off by default, increases program size due to formatting)
 * Daisy: set bootloader type in Makefile
+* Daisy: MIDI i/o for NoteOn/Off, ControlChange, ProcramChange, ChannelPressure, and PitchBend
 * DPF bugfixes: broken midi template include; MIDI_RT_CLOCK fails under certain conditions
 
 0.8.0